Break out nested classes in Application
Move the nested classes in Application to separate files.
